pyproject.toml
[project]
name = "personal-helper"
version = "0.1.0"
description = "MCP server exposing PLN→USD conversion, outfit picker, and Tavily search."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "mcp>=1.14.0",
  "python-dotenv>=1.0",
  "yfinance>=0.2.50",
  "tavily-python>=0.3.3",
  "fastapi>=0.117.1",
  "uvicorn>=0.36.0",
  "google-adk>=1.14.1",
  "google-cloud-aiplatform>=1.115.0",
]
